refactor: MemoState inner class 전부 제거
