+++ src/main/java/com/quickmemo/plugin/infrastructure/MemoState.java
@@ -1,69 +1,48 @@
 package com.quickmemo.plugin.infrastructure;
 
-import com.intellij.openapi.components.*;
+import com.intellij.openapi.components.PersistentStateComponent;
+import com.intellij.openapi.components.Service;
+import com.intellij.openapi.components.State;
+import com.intellij.openapi.components.Storage;
 import com.quickmemo.plugin.memo.Memo;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.*;
+import java.util.List;
 
 @Service(Service.Level.PROJECT)
 @State(
         name = "QuickMemoContent",
         storages = {@Storage("quickMemo.xml")}
 )
-public final class MemoState implements PersistentStateComponent<MemoState.State> {
-    private State state = new State();
-
-    public static class State {
-        public List<MemoLocalState> memos = new ArrayList<>();
-    }
-
-    public static class MemoLocalState {
-        public String id;
-        public String content;
-        public String createdAt;
-
-        public MemoLocalState() {
-        }
-
-        public MemoLocalState(String id, String content, String createdAt) {
-            this.id = id;
-            this.content = content;
-            this.createdAt = createdAt;
-        }
-    }
+public final class MemoState implements PersistentStateComponent<Memos> {
+    private Memos state = new Memos();
 
     public List<Memo> getAll() {
-        return state.memos.stream()
+        return state.memos.values().stream()
                 .map(data -> new Memo(data.id, data.content, data.createdAt))
                 .toList();
     }
 
     public void add(Memo memo) {
-        for (int i = 0; i < state.memos.size(); i++) {
-            if (state.memos.get(i).id.equals(memo.id())) {
-                return;
-            }
+        if (!state.memos.containsKey(memo.id())) {
+            state.memos.put(memo.id(), new MemoLocalState(memo.id(), memo.content(), memo.createdAt()));
         }
-        state.memos.add(new MemoLocalState(memo.id(), memo.content(), memo.createdAt()));
     }
 
     public void update(Memo memo) {
-        for (MemoLocalState data : state.memos) {
-            if (data.id.equals(memo.id())) {
-                data.content = memo.content();
-                break;
-            }
+        MemoLocalState data = state.memos.get(memo.id());
+        if (data != null) {
+            data.content = memo.content();
         }
     }
 
     public void remove(String id) {
-        state.memos.removeIf(data -> data.id.equals(id));
+        state.memos.remove(id);
     }
 
     public boolean isExist(String id) {
-        return state.memos.stream().anyMatch(data -> data.id.equals(id));
+        return state.memos.containsKey(id);
     }
 
     public boolean isEmpty() {
@@ -71,12 +50,12 @@ public final class MemoState implements PersistentStateComponent<MemoState.State
     }
 
     @Override
-    public @Nullable State getState() {
+    public @Nullable Memos getState() {
         return state;
     }
 
     @Override
-    public void loadState(@NotNull State state) {
+    public void loadState(@NotNull Memos state) {
         this.state = state;
     }
 }
