+++ src/main/java/com/quickmemo/plugin/infrastructure/Memos.java
@@ -1,0 +1,8 @@
+package com.quickmemo.plugin.infrastructure;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class Memos {
+    public Map<String, MemoLocalState> memos = new HashMap<>();
+}
